Use absolute offset for semantic scoping
Creating frames for semantic scopes such as loops and conditionals
has impact on performance. Relative offset computation is more
complicated and leads to higher risk of introducting bugs to the
compiler back-end.

Absolute offsets will be used for the main scope and activation
records for function calls. For the main scope the point of reference
is the start of the outer-most scope. Within each function the point
of reference is the functon entry as before.